requirements.txt
python-dotenv==0.17.1
selenium==4.0.0a1
webdriver-manager==4.0.1
openai==1.3.5

# beautifulsoup4==4.9.3
# bs4==0.0.1
# lxml==4.9.3
# requests==2.23.0